move Month enum to parsing
